fixed post-processing implementation
problem: T-FLEX CAD API cannot read the link by shortcut if the target fragment is in a subdirectory relative to the assembly file.
solution: if the target item contains at least 2 external links, then the link template is canceled.

+++ src/TFlex.PackageManager.UI/Views/Main.xaml.cs
@@ -1068,6 +1068,11 @@ namespace TFlex.PackageManager.UI.Views
                 if (item.Value.Flags != 1)
                     continue;
 
+                counter[0] += increment;
+                Marshal.Copy(counter, 0, value, counter.Length);
+                NativeMethods.SendMessage(handle, WM_INCREMENT_PROGRESS,
+                    IntPtr.Zero, value);
+
                 if (stoped)
                 {
                     NativeMethods.SendMessage(handle, WM_STOPPED_PROCESSING, 
@@ -1077,23 +1082,20 @@ namespace TFlex.PackageManager.UI.Views
                 }
 
                 proc.ProcessingFile(item.Value);
-
-                counter[0] += increment;
-                Marshal.Copy(counter, 0, value, counter.Length);
-                NativeMethods.SendMessage(handle, WM_INCREMENT_PROGRESS, 
-                    IntPtr.Zero, value);
             }
 
             foreach (var item in items)
             {
                 //
-                // processing the shortcut items
+                // post-processing items
                 //
                 if (item.Value.Flags != 4)
                     continue;
 
-                if (item.Value.ERefs.Count < 2)
-                    continue;
+                counter[0] += increment;
+                Marshal.Copy(counter, 0, value, counter.Length);
+                NativeMethods.SendMessage(handle, WM_INCREMENT_PROGRESS,
+                    IntPtr.Zero, value);
 
                 if (stoped)
                 {
@@ -1103,31 +1105,29 @@ namespace TFlex.PackageManager.UI.Views
                     break;
                 }
 
-                logging.WriteLine(LogLevel.INFO, "--- Processing the shortcut items");
+                if (item.Value.ERefs.Count < 2)
+                    continue;
+
+                logging.WriteLine(LogLevel.INFO, "--- Processing unshared resources");
 
                 item.Value.Flags ^= 0x5;
                 proc.ProcessingFile(item.Value);
 
                 foreach (var i in item.Value.ERefs)
                 {
-                    i.Flags ^= 0x1;
                     File.Delete(i.OPath);
                     logging.WriteLine(LogLevel.INFO, 
                         string.Format("0-7 Processing [path: {0}]", 
                         i.OPath));
-                    Helper.CreateShortcut(i.OPath, item.Value.OPath);
-                    logging.WriteLine(LogLevel.INFO,
-                        string.Format("0-8 Processing [path: {0}]",
-                        i.OPath));
-                    proc.ProcessingParent(i);
+
+                    var dir = Path.GetDirectoryName(i.OPath);
+                    if (Directory.GetFiles(dir).Length == 0)
+                        Directory.Delete(dir);
+
+                    proc.ProcessingParent(i, item.Value.OPath);
                 }
 
-                item.Value.ERefs.Clear();
-
-                counter[0] += increment;
-                Marshal.Copy(counter, 0, value, counter.Length);
-                NativeMethods.SendMessage(handle, WM_INCREMENT_PROGRESS,
-                    IntPtr.Zero, value);
+                item.Value.Flags ^= 0x5;
             }
 
             counter[0] = 100;
